Image decoders: harden PNG/JPG/GIF/Pamf input handling
* cellPngDec: pngDecReadBuffer now bounds the read against the declared
  buffer length and emits png_error if exhausted, and cellPngDecExtReadHeader
  pushes only the remaining bytes (length - cursor) instead of the full
  declared length, which previously walked past the guest buffer's end.
* cellJpgDec: cap input streamSize/fileSize at 64 MiB so a crafted size
  cannot drive multi-GB host allocations, fix the JFIF block-length parser
  (*0x100 instead of *0xFF), and bounds-check every byte read (i+1, i+8)
  before indexing into the buffer. Promote width*height*nComponents and
  per-line arithmetic in cellJpgDecDecodeData to u64 to stop int overflow
  when stbi returns large dimensions.
* cellGifDec: same input-size cap and width*height*nComponents promotion,
  and populate the subHandle out-pointer in cellGifDecExtOpen (it was
  previously left uninitialised so guest dereference produced garbage).
* cellPamf: compute ep_offset + ep_num * sizeof(PamfEpHeader) in u64 so
  attacker-controlled ep_num cannot wrap u32 past the header bound.

+++ rpcs3/Emu/Cell/Modules/cellPamf.cpp
@@ -545,7 +545,8 @@ error_code pamfVerify(vm::cptr<PamfHeader> pAddr, u64 fileSize, vm::ptr<CellPamf
 				return { CELL_PAMF_ERROR_INVALID_PAMF, "pamfVerify() failed: invalid ep_offset" };
 			}
 
-			if (ep_offset + ep_num * sizeof(PamfEpHeader) > header_size)
+			// Promote to u64 so attacker-controlled ep_offset/ep_num cannot wrap u32 past the header bound.
+			if (static_cast<u64>(ep_offset) + static_cast<u64>(ep_num) * sizeof(PamfEpHeader) > header_size)
 			{
 				return { CELL_PAMF_ERROR_INVALID_PAMF, "pamfVerify() failed: invalid ep_num" };
 			}
